refactor(redux): refactore store
