+++ packages/client/src/components/Loader/Loader.css
@@ -31,6 +31,7 @@
     left: 0;
     top: 0;
     bottom: 0;
+    z-index: 1000;
   }
 
   @keyframes loader {
