+++ packages/client/src/layouts/Root/Root.css
@@ -6,11 +6,7 @@
   width: 100%;
   margin: auto;
 
-
-
   @media (max-width: 730px) {
     width: 95%;
   }
 }
-
-
