+++ packages/client/src/store/features/app/typings.ts
@@ -1,0 +1,6 @@
+import { UserProfile } from '../../../app.typings';
+
+export interface AppState {
+  isLoading: boolean;
+  userProfile: UserProfile | null;
+}
