+++ packages/client/src/store/features/auth/typings.ts
@@ -1,0 +1,5 @@
+export interface AuthState {
+  isLoading: boolean;
+  isAuthenticated: boolean;
+  error: string;
+}
