+++ packages/client/src/utils/Http/httpClient.ts
@@ -3,7 +3,6 @@ import axios, { AxiosError } from 'axios';
 import { APIError } from '../../api/typings';
 import { apiHasError } from './../api-utils';
 
-
 // Check if error response body contains reason of error
 axios.interceptors.response.use(
   response => response,
